refactor(mcp): split intervals.py (1559 LOC) into 4 domain modules
Split the monolithic intervals handler into focused modules:
- intervals_sync.py: sync-week, sync-remote, backfill (3 handlers)
- intervals_events.py: delete, list, update, create-note (4 handlers)
- intervals_activities.py: get-details, get-intervals, get-streams (3 handlers)
- intervals_analysis.py: compare, apply-workout (2 handlers)

Original intervals.py becomes a ~40 LOC re-export shim preserving all
existing imports. MCP routing unchanged. Updated test patch target for
load_workout_descriptions to point to intervals_sync module.
